In a fit of insanity I replaced many lists with length-tagged lists.
The cool thing is that this ensures statically that, for example, the number of
received results is equal to the number of requested results, and that the
number of job results is equal to the number of labels in the job.  It also
guarantees that every job has at least one label.